Make the wakeup code saner.
Move wakeup code out of the backends and into the top-level manager.
This was needed to make it possible for the manager to manage
wakeup events.

In addition, we gain the efficiency of using eventfd for wakeups
on Linux.

--HG--
rename : src/System/Event/Internal.hsc => src/System/Event/Internal.hs