Get the KQueue code actually working, calloo callay!
Previously, the code was using the normal kevent system call, but
this led to great hilarity under OS X 10.6.

The hsc2hs command performs its ugly magic by compiling and running
a C program, but under 10.6 the platform default is to build 64-bit
C programs, while GHC still generates 32-bit Haskell programs.  For
ABIs that use pointers, such as struct kevent, this causes hsc2hs
to compute 64-bit offsets for pointers, while the ABI used by GHC
is 32-bit.  The result?  Bad craziness!

Fortunately, a switch to the kevent64 system call does the trick.

I've also wrapped the changes structure in an MVar, so that uses
of it that span multiple threads will not explode messily.