Add an extra pipe pair for wakeup messages.
The previous code was subject to the control pipe filling up on
platforms that didn't support eventfd, which would result in failed
control message deliveries.  This was easily provoked by running
the code in benchmarks/Simple.hs with any non-zero delay.

In this changeset, we introduce an extra pipe pair, which we use
purely for wakeup messages.  The event manager knows it can read
messages from this pipe in bulk, and ignore them all, and the writer
knows that it can ignore any EAGAIN or EWOULDBLOCK errors, since
the reader will be woken by previously queued wakeups.