Fix a nasty race condition spotted by Mark Lentczner.
In the interval between updating the map and telling the back end which
events to listen for, it was possible for another thread to sneak in and
give the back end a possibly clashing event mask, leading to inconsistency
between what the manager and back end thought they were waiting for.

We now wrap the emFds element in an MVar that we hold for the entire
duration of modifying both the manager and the back end.  This costs us a
little performance, but the inter-measurement variation is large enough that
it swamps any difference, so I conclude that it must not be significant.