Fix another race spotted by Mark Lentczner.
This time, the race lay in System.Event.Thread. We had been unregistering
the callback in a different thread than the one that was waking us, so that
thread could loop twice and try a putMVar more than once before we could
perform our single takeMVar. The result was a deadlocked event manager, and
hence a dead program.

The fix is to pass the FdRegistration to the callback, so it can unregister
itself immediately if need be.

The pong benchmark is still racy, but in a different way than before. Sigh.