Work around build problem by requiring at least base-4
This fixes a linking problem when compiling code under tests/ and
benchmarks/ that appears when Cabal selected base-3 when compiling
libHSevent.a.  An alternative fix would be to pass -package
base-3.x.x.x to GHC when linking, although this forces us to commit to
a certain version of base-3, which everyone might not have on their
system.

Better solutions welcome.